+++ src/main/java/com/sanaiclub/project/model/dto/RequiredStackDTO.java
@@ -8,11 +8,11 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class RequiredStackDTO {
-    private Long projectStackId; // 중복 제거용 id
-    private Long stackId;
+    private Integer projectStackId; // 중복 제거용 id
+    private Integer stackId;
     private String stackName;
     private String category;
     private Integer stackLevel;
     private Integer stackYear;
-    private Long projectId;
+    private Integer projectId;
 }
